Fix bijbelrooster bb tag
