+++ lib/view/bijbelrooster/BijbelroosterView.class.php
@@ -1,16 +1,15 @@
 <?php
-namespace CsrDelft\view;
+namespace CsrDelft\view\bijbelrooster;
 
-use CsrDelft\model\BijbelroosterModel;
 use CsrDelft\view\View;
 use PDOStatement;
 
 
 /**
  * BijbelroosterView.class.php
- * 
+ *
  * @author P.W.G. Brussee <brussee@live.nl>
- * 
+ *
  */
 class BijbelroosterView implements View {
 
@@ -55,29 +54,6 @@ class BijbelroosterView implements View {
 
 	public function view() {
 		echo '<h1>Bijbelrooster</h1><p>Hier vindt u het bijbelrooster der C.S.R.. Uw favoriete bijbelvertaling kunt u instellen bij uw <a href="/instellingen">instellingen</a>.</p>';
-		echo $this->getHtml();
-	}
-
-}
-
-class BijbelroosterBBView extends BijbelroosterView {
-
-	public function __construct($dagen) {
-		$dagen = (int) max((int) $dagen, 2);
-		$van = strtotime('-' . $dagen . ' days');
-		$tot = strtotime('+' . $dagen . ' days');
-		$rooster = BijbelroosterModel::instance()->getBijbelroosterTussen($van, $tot);
-		parent::__construct($rooster);
-	}
-
-	public function getHtml() {
-		$html = '<div class="bb-block mededeling-grotebalk">';
-		$html .= parent::getHtml();
-		$html .= '<div class="titel" style="float:right;position:relative;bottom:1.5em;right:10px"><a href="/bijbelrooster">Bijbelleesrooster</a></div></div>';
-		return $html;
-	}
-
-	public function view() {
 		echo $this->getHtml();
 	}
